LayoutTests/fast/forms/input-width-height-attributes-without-renderer-not-loaded-image.html

<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
<script>
jsTestIsAsync = true;
</script>
</head>
<body>
<input type="image" id="imageSlow" src="resources/image-slow.pl" style="display:none">
<div id="console"></div>
<script language="JavaScript">
    var e = document.getElementById("imageSlow");
    shouldBe('e.width', '0');
    shouldBe('e.height', '0');
    finishJSTest();
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
